+++ lib/arctic/vendor.rb
@@ -1,4 +1,5 @@
 require_relative 'vendor/dispersal_api'
+require_relative 'validation_api'
 require_relative 'logger'
 
 module Arctic